[WNMGDS-2471] Restrict normalization rules for list elements (#2610)
* Remove the gap for unstyled lists to avoid breaking other people's stuff

While we do want styled lists in text content to be uniform across products, we don't need to unnecessarily target other components that use lists and apply extra gap, padding, and margin. Maybe I'll also remove the margin in another commit. We were already removing the padding. This is for situations like the vertical nav where we do semantically want to use a list, but we want to apply our own styles. We've had reports of teams needing to apply different `flex-direction` values and remove `gap`. We don't really want to mess with these styles if we don't have to. While we knew we _would_ come into conflict, we hoped it'd be easy to for teams to overcome and move past. That was possibly a mistake.

* WIP: Don't apply the normalization declarations to unstyled lists

* Further refine the restriction of the list normalization styles

and update VRT refs. Note that the StepList styles did change, but it actually better matches what we had in v6 and removes an extra bit of 0.5em gap between steps that actually made them a little off-balance. I don't think we really intended to introduce that gap before.

* Restore a comment explaining the unstyled list rule

* These styles aren't necessary anymore

* Update docs a bit to reflect change

* Don't need these flex properties

* Column direction breaks footer

* Fix `ol:not([role])` taking precednce over `li > ol`

* Update VRT refs for normalized lists