[NO-TICKET] Fix possibility of null reference in Dropdown (#2635)
Treat `selectedItem` as nullable
i
The ding-dang loose types in Dropdown didn’t catch this. The type of `selectedItem` is `any`, which means it can be null, which means I can cause an error by referencing `selectedItem.value`. Hopefully it doesn’t happen in real life. This ongoing bug with the theme and version dropdown data in Gatsby is causing there to be no options in the select at build time, which is an abnormal situation that hopefully doesn’t happen in real applications. I’m going to make the fix on main and just plan to release it in v8