Refining types and seeing if it would work to not use Popover
I just discovered an issue that exists with our previous implementation of Dropdown as well where if you put a select inside a Dialog, the menu will always be contained by the dialog instead and cause it to create a scroll area rather than rendering outside the bounding box of the dialog item. I'm wondering if the popover code would have accounted for that, so I'm going to turn that back on and see if it works. The reason I removed it was because we already have styles for the dropdown menu, and they just seemed to make the styling more complex than necessary.