Reset the controlled story (successful experiment)
I had changed this controlled story to wrap it in a `form` element so I could test native html form support. It successfully passed the selected value through the GET request when the form was submitted! This is an improvement over what we had before, because we didn't support native html forms.