Because of a chicken-and-egg problem and for the sake of clarity...
I might need to just make an assumption that we're only going to receive one TextField as the children and nothing else. If I go find the textField child at first and discard the rest, it will allow me to both reference its props when calling `useComboBox` _and_ use the results of `useCombobox` to modify the textField. I could do something trickier where I pull it and then insert it again, but that would sacrifice clarity. For now, I think my next change will be to turn `transformTextFieldChild` into `getTextFieldChild`