[WNMGDS-2488] Autocomplete to react-aria (#2665)
* Starting to figure out what might need to change in Autocomplete

* Dropdown options now support arbitrary ReactNode content

and it's basically free. I'll remove the story change in the next commit

* Remove illustrative example of arbitrary option content

* Index will have to do for the section key

* Starting the translation layer

* WIP: Realizing that you can't have small Autocompletes

This is a limitation of the current Autocomplete that I wonder if we could fix. If you give a `size="medium"` or `size="small"` to the text field in an autocomplete, the dropdown menu won't be sized correctly

* Because of a chicken-and-egg problem and for the sake of clarity...

I might need to just make an assumption that we're only going to receive one TextField as the children and nothing else. If I go find the textField child at first and discard the rest, it will allow me to both reference its props when calling `useComboBox` _and_ use the results of `useCombobox` to modify the textField. I could do something trickier where I pull it and then insert it again, but that would sacrifice clarity. For now, I think my next change will be to turn `transformTextFieldChild` into `getTextFieldChild`

* Typing in the text field and filtering results is working

* `onChange` is hooked up now

* Avoid errors when selectedKey is null and maintain backwards compatability

* Fix warning about lack of a label, and clean up stuff

* Don't show the selected items in the autocomplete lists

* Match the behavior of the clear button with what it was before

* Loading and no-results behavior now matches, but I think it could be improved

Right now you can't really read these status messages with a screen reader. At least it doesn't announce them.

* Fix ids in the aria attibutes

* WIP: Working through tests

Getting some warnings, and I've got these weird `aria-hidden="true"` properties on the labels

* itemToString has an equivalent in react-aria, and it's textValue

* We don't supply this param anymore

* Fix key error in Dropdown unit tests

The group's index was the same as the value of one of the options.

* I don't think we really need these close buttons

We have the ESCAPE key, and I can't seem to reach these buttons with VoiceOver. Maybe if I were using the Popover, it would have a special mechanism for announcing these

* Fix confusing naming of label prop and remove snapshot of text field

Sarah is working on React/Preact snapshots in another branch, and Preact has a problem with snapshots of input values. A lot of this was noise anyway

* Debugging why keyboard interactions don't work in Preact

This goes back to the issue I made a PR for on the Adobe Spectrum repository

* Only use Preact in Storybook by a setting, since Autocomplete doesn't work in Preact

* Update VRT refs after switching storybook to react

There seemed to be one issue with tooltips in the Preact version. Not sure why it was shifted over to the left, but we'll want to look into that.