[WNMGDS-2488][WNMGDS-2489] Swap `Downshift` for `react-aria` (#2663)
* Make a dummy react-aria select to see the effect on bundle size

* Simplify by removing Button component

* Adding types after experimenting with a combined DropdownMenu component

* Add code from my first attempt at conversion a couple months ago

* Add some more example code from the docs

* Add support for sections

* Working on the typing for a combined popover-listbox component

It appears as though `listBoxRef` is in none of the type defs that are used in the example, and when I log the `menuProps` from `useSelect`, it doesn't exist there. I'm just going to remove it

* Can get rid of support for this prop that we don't use

* The combined popover+listbox component is working now

* Starting to create modules and apply CSS classes

* Refining types and seeing if it would work to not use Popover

I just discovered an issue that exists with our previous implementation of Dropdown as well where if you put a select inside a Dialog, the menu will always be contained by the dialog instead and cause it to create a scroll area rather than rendering outside the bounding box of the dialog item. I'm wondering if the popover code would have accounted for that, so I'm going to turn that back on and see if it works. The reason I removed it was because we already have styles for the dropdown menu, and they just seemed to make the styling more complex than necessary.

* Tried to use Popover to fix dropdown-in-dialog problem, but didn't work

I uncommented the code that put the listbox inside a Popover, which renders the listbox inside a portal which just appends it to the document body, but that means it renders under the HTML `dialog`, which is an even worse problem than our original problem.

* Came up with a fix for dropdowns in dialogs but did not implement

This would require a non-trivial change to how structure our dialog styles, but it's essetially making the dialog element itself take up the full screen but having a child element act as the "window" so it doesn't look any different.

* Working on applying everything to the Dropdown component itself

* Use react-stately's setFocus method and make a TODO

* Use our own ids for everything

* Add react-aria's `HiddenSelect`

* Pass real option data through and fix the group styles

What this doesn't do yet is pass arbitrary html attributes from `<option>` down to the `DropdownMenuOption`

* Support value and defaultValue

* Extra attributes for `<option>` and `<optgroup>` are passed down!

* Delete prototyping code

* Add support for uncontrolled components and fix default value bug

It couldn't actually find a value in option groups

* Use the same onBlur calculation strategy as we do in date field

I think this strategy has room for improvement to make it more robust, but for now it's working better than it was before. It doesn't fire when it's not supposed to.

* Clean up commented-out code

* Don't need the svg icon title because it's aria-hidden

* Reset the controlled story (successful experiment)

I had changed this controlled story to wrap it in a `form` element so I could test native html form support. It successfully passed the selected value through the GET request when the form was submitted! This is an improvement over what we had before, because we didn't support native html forms.

* Remove to-do comment (complete!)

* Tidy up some code

* Fix unit tests and add another

* Add tests for passing arbitrary attrs and allow users to override group label ids

* Remove the dialog story (will add in separate bugfix)

* Support class names for Autocomplete

* Fix interaction tests

* Restore the focus ring when the dropdown is open

Because we actually move focus between the options, the button technically isn't in a focus state. To match our other components better and to keep the current styling, I've added come CSS to make the focus ring show

* Correctly pass disabled state to react-aria

* Angela says this might be better as a combobox because it's more familiar

* Tabbing selects the currently highlighted option

Right now it's firing two blur events. Need to figure out how to prevent that in a way that doesn't break something else. I've attempted to fix it, and it's only caused more problems so far

* Fix the double blur event

* Flip flop flip flop

* [WNMGDS-2488] Autocomplete to react-aria (#2665)

* Starting to figure out what might need to change in Autocomplete

* Dropdown options now support arbitrary ReactNode content

and it's basically free. I'll remove the story change in the next commit

* Remove illustrative example of arbitrary option content

* Index will have to do for the section key

* Starting the translation layer

* WIP: Realizing that you can't have small Autocompletes

This is a limitation of the current Autocomplete that I wonder if we could fix. If you give a `size="medium"` or `size="small"` to the text field in an autocomplete, the dropdown menu won't be sized correctly

* Because of a chicken-and-egg problem and for the sake of clarity...

I might need to just make an assumption that we're only going to receive one TextField as the children and nothing else. If I go find the textField child at first and discard the rest, it will allow me to both reference its props when calling `useComboBox` _and_ use the results of `useCombobox` to modify the textField. I could do something trickier where I pull it and then insert it again, but that would sacrifice clarity. For now, I think my next change will be to turn `transformTextFieldChild` into `getTextFieldChild`

* Typing in the text field and filtering results is working

* `onChange` is hooked up now

* Avoid errors when selectedKey is null and maintain backwards compatability

* Fix warning about lack of a label, and clean up stuff

* Don't show the selected items in the autocomplete lists

* Match the behavior of the clear button with what it was before

* Loading and no-results behavior now matches, but I think it could be improved

Right now you can't really read these status messages with a screen reader. At least it doesn't announce them.

* Fix ids in the aria attibutes

* WIP: Working through tests

Getting some warnings, and I've got these weird `aria-hidden="true"` properties on the labels

* itemToString has an equivalent in react-aria, and it's textValue

* We don't supply this param anymore

* Fix key error in Dropdown unit tests

The group's index was the same as the value of one of the options.

* I don't think we really need these close buttons

We have the ESCAPE key, and I can't seem to reach these buttons with VoiceOver. Maybe if I were using the Popover, it would have a special mechanism for announcing these

* Fix confusing naming of label prop and remove snapshot of text field

Sarah is working on React/Preact snapshots in another branch, and Preact has a problem with snapshots of input values. A lot of this was noise anyway

* Debugging why keyboard interactions don't work in Preact

This goes back to the issue I made a PR for on the Adobe Spectrum repository

* Only use Preact in Storybook by a setting, since Autocomplete doesn't work in Preact

* Update VRT refs after switching storybook to react

There seemed to be one issue with tooltips in the Preact version. Not sure why it was shifted over to the left, but we'll want to look into that.

* Restore ESC behavior of old Autocomplete

where if you press ESC it just closes the listbox. Before this fix, it was also clearing the input value