This might be a step back, but it seems to behave a little more predictably
with this simpler implementation. This is a fork of `preact-custom-elements`. Not sure we should keep it, but I wanted to save this version of the code

The problem I'm seeing right now in Storybook is that it doesn't appear to actually be registering the custom component. That could be what has been happening all this time. I've just learned how to check for it: https://stackoverflow.com/a/28210364