Avoid duplicate content inside web component
Adding a key forces the web component to rerender when the args change. Before I did this, it would render the alert body but outside the `<div class="ds-c-alert"` in the inner html of the `<ds-alert>` it would duplicate the original text content that was the children of the element.