Back to the original style of fix but with improvements
Rather than try to force it to open at certain times with an unreliable imperative call to the "open" function, embrace the library's own logic for determining the `isOpen` state but allow it to show empty collections (which [lets us pass through this logic gate](https://github.com/adobe/react-spectrum/blob/main/packages/%40react-stately/combobox/src/useComboBoxState.ts#L116-L119)). The problem with showing empty collections before was that we sometimes didn't have anything to show. That's why I've added another qualifier in the logic that determines whether we should render the dropdown menu to only rely on the value of `isOpen` if we have actual result items to show.

When testing it out in one of the applications, I encountered an error that was due to the `popoverRef.current` being undefined, probably due to either fact that I don't use _their_ popover and or to [the chicken/egg problem described in this react-spectrum issue](https://github.com/adobe/react-spectrum/issues/4016#issuecomment-1417369731). I fixed it by filtering out `undefined` values in one of the Adobe Spectrum functions that assumed that ref had a `.current` value.