Also test that it closes the menu after selecting
While this doesn't catch the problem with the [previous fix](https://github.com/CMSgov/design-system/pull/2694), it adds a good sanity check. I can't really test for the problem that it _didn't_ catch, because it involves passing a new set of `items` (only one item) after a selection is made, which assumes a pretty specific implementation on the application side.