[WNMGDS-2518] Fix additional Autocomplete open/close issues (#2705)
* WIP: Less naive item comparison, but it opens when it shouldn't

Specifically, if selecting an item narrows down the items array to the one that matches that item, it will open the menu right after selection, because the items changed

* Back to the original style of fix but with improvements

Rather than try to force it to open at certain times with an unreliable imperative call to the "open" function, embrace the library's own logic for determining the `isOpen` state but allow it to show empty collections (which [lets us pass through this logic gate](https://github.com/adobe/react-spectrum/blob/main/packages/%40react-stately/combobox/src/useComboBoxState.ts#L116-L119)). The problem with showing empty collections before was that we sometimes didn't have anything to show. That's why I've added another qualifier in the logic that determines whether we should render the dropdown menu to only rely on the value of `isOpen` if we have actual result items to show.

When testing it out in one of the applications, I encountered an error that was due to the `popoverRef.current` being undefined, probably due to either fact that I don't use _their_ popover and or to [the chicken/egg problem described in this react-spectrum issue](https://github.com/adobe/react-spectrum/issues/4016#issuecomment-1417369731). I fixed it by filtering out `undefined` values in one of the Adobe Spectrum functions that assumed that ref had a `.current` value.

* TESTING: Test this change in Storybook

* Revert "TESTING: Test this change in Storybook"

This reverts commit b8165c3e374229b67befc0e339692b5a87e96f58.

* Remove commented-out code

* Also test that it closes the menu after selecting

While this doesn't catch the problem with the [previous fix](https://github.com/CMSgov/design-system/pull/2694), it adds a good sanity check. I can't really test for the problem that it _didn't_ catch, because it involves passing a new set of `items` (only one item) after a selection is made, which assumes a pretty specific implementation on the application side.