[WNMGDS-2498] Add web-component support to our Storybook (#2715)
* Trying to get our web components working in our storybook

* Actually use Preact in storybook

* These imports need to be different. See comment where it's used

* WIP: Trying to figure out why dropdown children are wrong and figure out a way to use the src in storybook rather than dist

* Learning a lot but haven't fixed anything

* This might be a step back, but it seems to behave a little more predictably

with this simpler implementation. This is a fork of `preact-custom-elements`. Not sure we should keep it, but I wanted to save this version of the code

The problem I'm seeing right now in Storybook is that it doesn't appear to actually be registering the custom component. That could be what has been happening all this time. I've just learned how to check for it: https://stackoverflow.com/a/28210364

* Welp, I've got a Heisenbug. Works in storybook if I console.log this value

If I don't log that value, the custom element is not defined. If I do, it's defined. It's defined when I look at the iframe and not the wrapper window for storybook

* Back to `preactement` for now

* Avoid duplicate content inside web component

Adding a key forces the web component to rerender when the args change. Before I did this, it would render the alert body but outside the `<div class="ds-c-alert"` in the inner html of the `<ds-alert>` it would duplicate the original text content that was the children of the element.

* Add a local version of preactement so we can modify it

* WIP: Add property definition

It seems to work to be able to read `element.variation` and `element.heading` on our `ds-alert` story, but the problem is that the component children no longer renders

* Getting rid of this code brings the children back

I don't know if anything can be lost between when the constructor is called and when the component is mounted though

* 'default' didn't work for a variation in the storybook args

* Add a button story with a click listener that prints to the actions tab!

* Add `is-alternate` control to `ds-button` story

* Fixed the print-garbage-to-get-styles Heisenbug

The symptom was that I could only get the web component stories working (particularly their styles) if I imported garbage (there are no exports from these modules) and console logged it:

```tsx
// In ds-button.stories.tsx
import DsButton from './ds-button';
console.log(DsButton);
```

What printing out that value was doing was telling Storybook’s tree-shaking mechanism (webpack) to care about the effects of that module. I found [this storybook issue discussion](https://github.com/storybookjs/storybook/issues/14953) where it’s pointed out that the user’s `package.json` said `"sideEffects": "false"` when they actually should have listed their scss modules as side effects. I went to our `package.json` and saw that the only side effects I had listed were in the `dist` folder:

```json
"sideEffects": [
    "./dist/preact-components/esm/web-components/*.js"
],
```

Changed it to

```json
"sideEffects": [
    "./src/components/web-components/*",
    "./dist/preact-components/esm/web-components/*.js"
],
```

and I removed the weird import and console log, and it worked

* Some more story controls

* Don't need this anymore. Fixed importing directly from web components modules

* Don't need any of these changes anymore

* Remove the unfinished `ds-dropdown`

* Add a readme for the preactement fork code

* Don't need this babel config in storybook

* I can't get this globbing to work anyway. See all previous related commit messages

* Ignore the web component stories in VRT for now

until we switch to using Preact again for our tests

* Remove this unrelated typedef change

* Ignore the preactement code for our eslint checks