[WNMGDS-2525] Add a `ds-usa-banner` story (#2727)
* Add a `ds-usa-banner` story

Right now the id doesn't really work. It doesn't get passed down as a prop to the underlying Preact component, and we haven't really determined how ids should work on web components yet. To test the class-name attribute, give it something like `ds-u-border--2`

* The component re-renders with the id in the rendered content with this hack

Without that hack, the component content re-renders just fine when the `class-name` is changed but not when the `id` is changed.

* Use a wrapper to map the id prop from another name

The problem with using `id` and passing it down is that the outer custom element itself will also have that `id` attribute, which means there will be two elements on the page with the same `id` attribute, which is a no-no

* Use JSX syntax for now because it's a little more compact

* Revert changes to ds-button to reduce scope of pr

* Again, ignore the VRTs for now

* Typo

* Remove `class-name` attribute and update tests

There's nothing really gained by allowing users to place custom classes on the inner root element when they can place them on the outer custom element, and keeping it around just increases the surface area of our API

* Forgot to remove `class-name` from the story