Clean up and explanation
The workaround I found on the Adobe Spectrum repository doesn't work all the time as-is. If you are scrolling the dropdown menu on a touch device first and then lift up your finger and tap the item under your finger, the touch event will fire on the element behind the menu. It works if you don't scroll the menu, though. My change is to remove the `once: true` part. I think it was probably added in there for to improve performance, but I'm less concerned about performance than I am about accidentally clicking things we don't mean to click!