Debugging
I'm looking at the section with the id `hello-14` where in the healthcare theme it should be rendering as not hidden. In the console, it prints `👍 hello-14`, which means that it's rendering without the `hidden` attribute, and yet in the browser the element still has that attribute. It's not hydrating correctly!!!