Change the order of sections to lessen the effect of wild hydration
This doesn't fix the root issue, and it moves it around to a different element, but at least it's readable now.

In the long-term, we should probably go through our content and see if we can consolidate and get rid of the need for differences in the guidance. Sarah even had the idea to call out the differences by putting them in alerts so users understand that it’s an amendment to the core guidance.