[WNMGDS-2530] Sketch plugin: Update swatches in place (#2736)
* Add support for getting colors from components files as well

* Instead of replacing all swatches, selectively replace updated ones and add new ones

* Embed the token data in the sketch plugin and select theme from list instead of filesystem

Also updated the naming of the exported json tokens to not all be prefixed with "theme"

* [NO-TICKET] Expand token support in the sketch plugin to component colors (#2729)

* Add support for getting colors from components files as well

* Instead of replacing all swatches, selectively replace updated ones and add new ones

* Updates the swatches in place and updates references, but it only works once

Here's how I'm testing: I have a rectangle whose fill references the alert background token and whose border color references the alert border color token. When I switch themes, the rectangle changes colors. However, it only works once. I think this might have to do with changes not being saved

* If I also search and replace, it updates swatch references every time

However, if I only search and replace and do not use `updateWithColor` and `updateReferencesToSwatch` but simply replace the swatch in my swatch map with a new swatch, it does not work

* Cleaning up and organizing

Trying to figure out why this code works. It throws errors, and yet the loop for updating swatches in the layer styles is necessary for everything to work

* Found that the error keeps the rest of the function from executing

The error we get from `style.fillType === Swatch` causes the rest of the function not to execute. Apparently this only works if we don't update the `doc.swatches`!

* Determined that calls to `sketch.Swatch.from` do not automatically add to doc

It had appeared that way in my last commit, but I think the reason why I couldn't just replace the array of swatches with one generated from `Object.values(swatchMap)` was that Sketch was losing all its references to the swatches that I had mutated through `updateWithColor` when I replaced the array with a new instance. I don't know why that would happen, but my code seems to work if I don't reassign the whole array and instead use `.push` to add my new swatches if they're new

* Update plugin readme