[WNMGDS-2546] Add `ds-dropdown` web component (#2772)
* Add code to create the `ds-dropdown` web component, though it's broken

* Debugging dropdown children

I've been comparing the children received by the Alert and the Dropdown. Alert seems to always have non-empty React/Preact elements for its children, whether we look in the component, component wrapper, or the rendering `finaliseComponent` function in `preactement`. The Dropdown has a React/Preact element, but it doesn't appear to have any data whatsoever—no props, no children, nothing distinguishable. I can't tell what's different between the two, either in the component code, the custom-element definition module, or the Storybook stories.

* Update some props and argTypes

* Allow setting options via `element.options = ` or a stringified attribute

I still want to get the `children` option working, but this is one option that works for now

* Silence TS errors

* Get rid of warnings of having options and props at the same time

Scratch my previous comment about it. This is a simple enough fix

* Get rid of debugging logs

* Add ds-dropdown to the cdn example, get events working, and fix name duplication

There is a hidden `select` element that takes the `name` of the dropdown, and we want that to be the only one in the form with that name. We were accidentally adding the name to the dropdown button as well

* Adding support for aria-disabled, but we don't actually support it in the css

* Add a whole bunch of documentation, include an events table

* Refining how the event object details are presented in the table

* More attribute documentation

* Update snapshot for removing name from dropdown button

* The start of some unit tests

* Finish unit tests

* Fixed TS errors

* Don't take VRT snapshots of ds-dropdown

* Add TODO comment

* Remove duplicate unit test