The fix
Implements [this workaround](https://github.com/facebook/react/issues/11538#issuecomment-390386520) for React crashing when Google Translate manipulates the DOM. This is a difficult fix to very because it doesn't seem to fix it in Storybook. For a long time I thought that this fix wasn't enough because I couldn't get the errors to go away in Storybook, even when compiling a static version. I think this is good enough given that it does fix it in at least one environment. I've tried stripping away almost everything from the header and making the button only have an `onClick` and plain text for all its content, but that still didn't take away the error in Storybook. (I can only reproduce the error in storybook by opening the controlled logged-in story, turning on Google Translate, and then clicking the menu button. It doesn't even matter if the menu is rendered or not.)

I will commit some changes that allow us to test it in our `examples/create-react-app-typescript` project and then remove the changes.