Text styles now inherit "parent" text styles
I'm on the fence about also representing the inheritance structure in the style names. I might remove that part. It doesn't really make sense that the `typography-heading-2xl` style is directly in the `typography-heading` folder, but `typography-heading-2xl--mobile` is under a `typography-heading-2xl` folder that is a sibling of the `typography-heading-2xl` style. The alternative is to put every style in its own folder named itself, but that would be annoying for designers to have to navigate