[NO-TICKET] Fix the release notes script using older tags for version numbers (#2796)
Fix the release notes script using older tags

The problem with the command `git tag -l` was that it sorts tags by string value. Our logic relied on the sort order being numeric. After version 10, the most recent version no longer appeared at the end of the list. 9.0.2 is the last version in the list, because 10, 11, and 12 come before 9 in alphabetical string-sort order