[WNMGDS-2553] Sketch plugin: Add inheritance logic for text styles (#2783)
* Text styles now inherit "parent" text styles

I'm on the fence about also representing the inheritance structure in the style names. I might remove that part. It doesn't really make sense that the `typography-heading-2xl` style is directly in the `typography-heading` folder, but `typography-heading-2xl--mobile` is under a `typography-heading-2xl` folder that is a sibling of the `typography-heading-2xl` style. The alternative is to put every style in its own folder named itself, but that would be annoying for designers to have to navigate

* Revert folder name structure for reasons outlined in previous commit

* Avoid NaN and correctly handle 'inherit' and other unparseable values

* Clean up debugging output

* Add more comments

* Get Sketch to stop adding a random border to text styles

It would always add a #979797 1px border to all the text styles, and I could not figure out why

* Replace two `replaceAll` calls with one global regex replace