[WNMGDS-2571] Combine browser tests for faster CI (#2806)
* Combine a11y tests into other browser tests to speed them up

Right now it takes CI 6m37s to run our a11y tests. Combining them with the smoke VRT tests means we can not only run fewer of them, but we can load each story once and run two different tests against it. It takes CI way too long to load these pages, so this should hopefully speed things up. It takes our CI 3m47s to run the VRT smoke tests. That's the number to compare to.

* Forgot to add the browser-name logic

* Don't load the page again! That's the whole point!

* Temporarily skip a test that is throwing an error

We were testing that story before but not on a screen small enough to make the problematic element appear.

1. Go [to this story](https://design.cms.gov/storybook/iframe.html?args=&id=components-table--scrollable-table&viewMode=story "Follow link")
2. Resize the screen until it’s small enough for a blue alert to show up at the top
3. Run Axe core and look at the error (see below and attached screenshot)

* Close the context that we created

* Exclude forced colors mode, which is run in chromium