Fix the problem by replacing `isolation` with `position: relative`
And because of the relative positioning, the skipnav also need to be repositioned. Instead of trying to shoehorn it into the right position, I moved it to be the first child of the header component in the DOM, since that is actually where it's supposed to be anyway [according to the USWDS guidance](https://designsystem.digital.gov/components/banner/) (skipnav supposed to be right before the banner).