Discovery: Using a boolean prop fixes the bug
I've temporarily refactored `NativeDialog`, `Dialog`, and the dialog stories to use a boolean prop on `Dialog` to toggle the open state rather than conditional rendering of the `Dialog`. Running storybook with `yarn storybook:react`, I'm able to reproduce the bug when closing the dialog with the close button, and after the refactoring, the bug goes away. So there's something here.

Same thing happens when I do the same to the `Drawer` and drawer stories. The bug goes away.

I don't really want to change the API, so how can I figure out what part of this is relevant and apply it to the old API?