WIP: Found a bug in the `NativeDialog` `useEffect` function
When I ran this in the `DrawerManager` story (which has multiple dialogs), I found that when I opened a new drawer when another one was open, the dialogs were closing only in the cleanup functions. Then I realized that the cleanup functions were being called on dialogs that should be open, which was a silly mistake of mine. I'm always closing in the cleanup function when, in fact, it will execute the cleanup function for more reasons than just de-rendering the component. It can also execute the cleanup function just because isOpen changed!