[WNMGDS-2555] Add `isOpen` to `Drawer` and `Dialog` to fix focus-management issue (#2890)
* Discovery: Using a boolean prop fixes the bug

I've temporarily refactored `NativeDialog`, `Dialog`, and the dialog stories to use a boolean prop on `Dialog` to toggle the open state rather than conditional rendering of the `Dialog`. Running storybook with `yarn storybook:react`, I'm able to reproduce the bug when closing the dialog with the close button, and after the refactoring, the bug goes away. So there's something here.

Same thing happens when I do the same to the `Drawer` and drawer stories. The bug goes away.

I don't really want to change the API, so how can I figure out what part of this is relevant and apply it to the old API?

* Finish the temporary scroll code

Previously when I was prototyping this, I just replaced the scroll code with garbage so it'd compile. I don't love that not having the effect cleanup function closure means I have to track the `y` state separately, but whatever

* What about a finite state machine?

* Leverage TS a bit more in the finite state machine

* Clean up the property when we're done

* Add warning for old way of using these components

* Create some symmetry in the code blocks to make it easier to understand

* Forgot the temprary backwards compatability part

* Add unit tests for Dialog

* Add unit tests for Drawer

* Make the Dialog test module TypeScript

* The unreleased DrawerManager stories needed to be updated too

And I actually think we can release these now that this bug is fixed

* Go update all the places where we use dialogs

* Update a bunch of tests, but need to work on DrawerManager tests

* Make this prop optional for now until deprecation period is over

* WIP: Found a bug in the `NativeDialog` `useEffect` function

When I ran this in the `DrawerManager` story (which has multiple dialogs), I found that when I opened a new drawer when another one was open, the dialogs were closing only in the cleanup functions. Then I realized that the cleanup functions were being called on dialogs that should be open, which was a silly mistake of mine. I'm always closing in the cleanup function when, in fact, it will execute the cleanup function for more reasons than just de-rendering the component. It can also execute the cleanup function just because isOpen changed!

* We don't actually need to close the dialog in the effect cleanup (fixes bug)

I originally thought that we needed to make sure it closes if the component ever completely de-renders and is removed from the DOM for good, but removing it from the DOM will also close the dialog, just without the callbacks for close, which we don't want anyway

* Fix calling close handler when we don't want to

* This DrawerManager unit tests might not work

It might not work just because of how jsdom handles dialogs. That's my theory anyway. I think it just doesn't have the focus-managing functionality

* See previous commit. Don't think this is ever going to work in jsdom

* Don't need this prop in this test

* Remove debugging output

* Fix TS error

* Update storybook docs snapshots

* Fix drawer toggle story so it doesn't start open